fixes in repsonse to comments
